Backend/MachineLearningAPI/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
click==8.2.1
fastapi==0.115.13
frozenlist==1.7.0
h11==0.16.0
idna==3.10
joblib==1.5.1
multidict==6.5.1
numpy==2.3.1
pandas==2.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
xgboost==3.0.2
yarl==1.20.1
